Kconfig: introduce LEGACY_PCI config option
Introduce a new LEGACY_PCI config option which enables legacy PCI
devices including those attached to a PCI-to-PCI Express bridge and PCI
Express devices using legacy I/O spaces. Note that this is different
from non PCI uses of I/O ports such as by ACPI.

The option defaults to y and subsequent patches will add it as
a dependency for PCI drivers on a per subsystem basis.

Note: The use of "legacy" above matches the wording of the PCIe
specification when referring to the I/O spaces compatibility feature.

Co-developed-by: Arnd Bergmann <arnd@kernel.org>
Signed-off-by: Arnd Bergmann <arnd@kernel.org>
Signed-off-by: Niklas Schnelle <schnelle@linux.ibm.com>